Move tests from `smoke-tests/validation-rules` to `base`.
