+++ base/src/test/java/io/spine/validate/AbstractValidatingBuilderShould.java
