+++ base/src/test/java/io/spine/validate/ConstraintViolationsShould.java
@@ -20,6 +20,8 @@
 package io.spine.validate;
 
 import com.google.common.testing.NullPointerTester;
+import io.spine.validate.ConstraintViolation;
+import io.spine.validate.ConstraintViolations;
 import org.junit.Test;
 
 import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
