+++ base/src/test/java/io/spine/validate/FieldValidatorFactoryShould.java
@@ -29,17 +29,17 @@ import com.google.protobuf.FloatValue;
 import com.google.protobuf.Int32Value;
 import com.google.protobuf.Int64Value;
 import com.google.protobuf.StringValue;
-import io.spine.test.validate.msg.MessageWithMapBoolField;
-import io.spine.test.validate.msg.MessageWithMapByteStringField;
-import io.spine.test.validate.msg.MessageWithMapDoubleField;
-import io.spine.test.validate.msg.MessageWithMapFloatField;
-import io.spine.test.validate.msg.MessageWithMapIntField;
-import io.spine.test.validate.msg.MessageWithMapLongField;
-import io.spine.test.validate.msg.MessageWithMapMessageField;
-import io.spine.test.validate.msg.MessageWithMapStringField;
-import io.spine.test.validate.msg.RequiredByteStringFieldValue;
-import io.spine.test.validate.msg.RequiredEnumFieldValue;
-import io.spine.test.validate.msg.RequiredMsgFieldValue;
+import io.spine.test.validate.MessageWithMapBoolField;
+import io.spine.test.validate.MessageWithMapByteStringField;
+import io.spine.test.validate.MessageWithMapDoubleField;
+import io.spine.test.validate.MessageWithMapFloatField;
+import io.spine.test.validate.MessageWithMapIntField;
+import io.spine.test.validate.MessageWithMapLongField;
+import io.spine.test.validate.MessageWithMapMessageField;
+import io.spine.test.validate.MessageWithMapStringField;
+import io.spine.test.validate.RequiredByteStringFieldValue;
+import io.spine.test.validate.RequiredEnumFieldValue;
+import io.spine.test.validate.RequiredMsgFieldValue;
 import org.junit.Test;
 
 import static com.google.common.collect.ImmutableMap.of;
