+++ base/src/test/java/io/spine/validate/IntegerFieldValidatorShould.java
