+++ base/src/test/java/io/spine/validate/MessageValidatorShould.java
@@ -32,48 +32,47 @@ import io.spine.base.FieldPath;
 import io.spine.option.OptionsProto;
 import io.spine.option.Time;
 import io.spine.protobuf.AnyPacker;
-import io.spine.test.validate.msg.CustomMessageRequiredByteStringFieldValue;
-import io.spine.test.validate.msg.CustomMessageRequiredEnumFieldValue;
-import io.spine.test.validate.msg.CustomMessageRequiredMsgFieldValue;
-import io.spine.test.validate.msg.CustomMessageRequiredRepeatedMsgFieldValue;
-import io.spine.test.validate.msg.CustomMessageRequiredStringFieldValue;
-import io.spine.test.validate.msg.CustomMessageWithNoRequiredOption;
-import io.spine.test.validate.msg.DecimalMaxIncNumberFieldValue;
-import io.spine.test.validate.msg.DecimalMaxNotIncNumberFieldValue;
-import io.spine.test.validate.msg.DecimalMinIncNumberFieldValue;
-import io.spine.test.validate.msg.DecimalMinNotIncNumberFieldValue;
-import io.spine.test.validate.msg.DigitsCountNumberFieldValue;
-import io.spine.test.validate.msg.EnclosedMessageFieldValue;
-import io.spine.test.validate.msg.EnclosedMessageFieldValueWithCustomInvalidMessage;
-import io.spine.test.validate.msg.EnclosedMessageFieldValueWithoutAnnotationFieldValueWithCustomInvalidMessage;
-import io.spine.test.validate.msg.EnclosedMessageWithRequiredString;
-import io.spine.test.validate.msg.EnclosedMessageWithoutAnnotationFieldValue;
-import io.spine.test.validate.msg.EntityIdByteStringFieldValue;
-import io.spine.test.validate.msg.EntityIdDoubleFieldValue;
-import io.spine.test.validate.msg.EntityIdIntFieldValue;
-import io.spine.test.validate.msg.EntityIdLongFieldValue;
-import io.spine.test.validate.msg.EntityIdMsgFieldValue;
-import io.spine.test.validate.msg.EntityIdRepeatedFieldValue;
-import io.spine.test.validate.msg.EntityIdStringFieldValue;
-import io.spine.test.validate.msg.FirstRuleTarget;
-import io.spine.test.validate.msg.InvalidMessage;
-import io.spine.test.validate.msg.MaxNumberFieldValue;
-import io.spine.test.validate.msg.MinNumberFieldValue;
-import io.spine.test.validate.msg.PatternStringFieldValue;
-import io.spine.test.validate.msg.RepeatedRequiredMsgFieldValue;
-import io.spine.test.validate.msg.RequiredBooleanFieldValue;
-import io.spine.test.validate.msg.RequiredByteStringFieldValue;
-import io.spine.test.validate.msg.RequiredEnumFieldValue;
-import io.spine.test.validate.msg.RequiredMsgFieldValue;
-import io.spine.test.validate.msg.RequiredStringFieldValue;
-import io.spine.test.validate.msg.RuleTargetAggregate;
-import io.spine.test.validate.msg.SecondRuleTarget;
-import io.spine.test.validate.msg.TimeInFutureFieldValue;
-import io.spine.test.validate.msg.TimeInPastFieldValue;
-import io.spine.test.validate.msg.TimeWithoutOptsFieldValue;
-import io.spine.test.validate.msg.altfields.AnyContainer;
-import io.spine.test.validate.msg.altfields.UncheckedAnyContainer;
-import io.spine.validate.given.MessageValidatorTestEnv;
+import io.spine.test.validate.CustomMessageRequiredByteStringFieldValue;
+import io.spine.test.validate.CustomMessageRequiredEnumFieldValue;
+import io.spine.test.validate.CustomMessageRequiredMsgFieldValue;
+import io.spine.test.validate.CustomMessageRequiredRepeatedMsgFieldValue;
+import io.spine.test.validate.CustomMessageRequiredStringFieldValue;
+import io.spine.test.validate.CustomMessageWithNoRequiredOption;
+import io.spine.test.validate.DecimalMaxIncNumberFieldValue;
+import io.spine.test.validate.DecimalMaxNotIncNumberFieldValue;
+import io.spine.test.validate.DecimalMinIncNumberFieldValue;
+import io.spine.test.validate.DecimalMinNotIncNumberFieldValue;
+import io.spine.test.validate.DigitsCountNumberFieldValue;
+import io.spine.test.validate.EnclosedMessageFieldValue;
+import io.spine.test.validate.EnclosedMessageFieldValueWithCustomInvalidMessage;
+import io.spine.test.validate.EnclosedMessageFieldValueWithoutAnnotationFieldValueWithCustomInvalidMessage;
+import io.spine.test.validate.EnclosedMessageWithRequiredString;
+import io.spine.test.validate.EnclosedMessageWithoutAnnotationFieldValue;
+import io.spine.test.validate.FirstRuleTarget;
+import io.spine.test.validate.InvalidMessage;
+import io.spine.test.validate.MaxNumberFieldValue;
+import io.spine.test.validate.MinNumberFieldValue;
+import io.spine.test.validate.PatternStringFieldValue;
+import io.spine.test.validate.RepeatedRequiredMsgFieldValue;
+import io.spine.test.validate.RequiredBooleanFieldValue;
+import io.spine.test.validate.RequiredByteStringFieldValue;
+import io.spine.test.validate.RequiredEnumFieldValue;
+import io.spine.test.validate.RequiredMsgFieldValue;
+import io.spine.test.validate.RequiredStringFieldValue;
+import io.spine.test.validate.RuleTargetAggregate;
+import io.spine.test.validate.SecondRuleTarget;
+import io.spine.test.validate.TimeInFutureFieldValue;
+import io.spine.test.validate.TimeInPastFieldValue;
+import io.spine.test.validate.TimeWithoutOptsFieldValue;
+import io.spine.test.validate.anyfields.AnyContainer;
+import io.spine.test.validate.anyfields.UncheckedAnyContainer;
+import io.spine.test.validate.commands.EntityIdByteStringFieldValue;
+import io.spine.test.validate.commands.EntityIdDoubleFieldValue;
+import io.spine.test.validate.commands.EntityIdIntFieldValue;
+import io.spine.test.validate.commands.EntityIdLongFieldValue;
+import io.spine.test.validate.commands.EntityIdMsgFieldValue;
+import io.spine.test.validate.commands.EntityIdRepeatedFieldValue;
+import io.spine.test.validate.commands.EntityIdStringFieldValue;
 import org.junit.Test;
 
 import java.util.List;
@@ -85,12 +84,12 @@ import static com.google.protobuf.util.Timestamps.subtract;
 import static io.spine.base.Identifier.newUuid;
 import static io.spine.base.Time.getCurrentTime;
 import static io.spine.protobuf.TypeConverter.toMessage;
-import static io.spine.validate.given.MessageValidatorTestEnv.FIFTY_NANOSECONDS;
-import static io.spine.validate.given.MessageValidatorTestEnv.SECONDS_IN_5_MINUTES;
-import static io.spine.validate.given.MessageValidatorTestEnv.ZERO_NANOSECONDS;
-import static io.spine.validate.given.MessageValidatorTestEnv.currentTimeWithNanos;
-import static io.spine.validate.given.MessageValidatorTestEnv.freezeTime;
-import static io.spine.validate.given.MessageValidatorTestEnv.timeWithNanos;
+import static io.spine.validate.MessageValidatorTestEnv.FIFTY_NANOSECONDS;
+import static io.spine.validate.MessageValidatorTestEnv.SECONDS_IN_5_MINUTES;
+import static io.spine.validate.MessageValidatorTestEnv.ZERO_NANOSECONDS;
+import static io.spine.validate.MessageValidatorTestEnv.currentTimeWithNanos;
+import static io.spine.validate.MessageValidatorTestEnv.freezeTime;
+import static io.spine.validate.MessageValidatorTestEnv.timeWithNanos;
 import static java.lang.String.format;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
