+++ base/src/test/java/io/spine/validate/StringFieldValidatorShould.java
@@ -1,11 +1,31 @@
+/*
+ * Copyright 2018, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
 package io.spine.validate;
 
 import com.google.protobuf.Descriptors.FieldDescriptor;
-import io.spine.test.validate.msg.MessageWithMapStringField;
-import io.spine.test.validate.msg.MessageWithRepeatedRequiredValidatedStringField;
-import io.spine.test.validate.msg.MessageWithRepeatedUnchekedStringField;
-import io.spine.test.validate.msg.MessageWithRepeatedValidatedStringField;
-import io.spine.test.validate.msg.MessegeWithRepeatedRequiredStringField;
+import io.spine.test.validate.MessageWithMapStringField;
+import io.spine.test.validate.MessageWithRepeatedRequiredValidatedStringField;
+import io.spine.test.validate.MessageWithRepeatedUnchekedStringField;
+import io.spine.test.validate.MessageWithRepeatedValidatedStringField;
+import io.spine.test.validate.MessegeWithRepeatedRequiredStringField;
 
 import java.util.Collections;
 import java.util.List;
