+++ base/src/test/proto/spine/test/validate/any_fields.proto
@@ -19,46 +19,23 @@
 //
 syntax = "proto3";
 
-package spine.test.validate.msg;
+package spine.test.validate;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
 option java_multiple_files = true;
-option java_outer_classname = "ValidationTestCommandsProto";
-option java_package = "io.spine.test.validate.msg";
+option java_outer_classname = "ValidationTestAnyFieldsProto";
+option java_package = "io.spine.test.validate.anyfields";
 
-import "google/protobuf/wrappers.proto";
+import "google/protobuf/any.proto";
 
+message AnyContainer {
 
-// Messages for entity ID in commands validation tests.
-
-message EntityIdMsgFieldValue {
-    google.protobuf.StringValue value = 1;
+    google.protobuf.Any any = 1 [(valid) = true];
 }
 
-message EntityIdStringFieldValue {
-    string value = 1;
-}
+message UncheckedAnyContainer  {
 
-message EntityIdIntFieldValue {
-    int32 value = 1;
-}
-
-message EntityIdLongFieldValue {
-    int64 value = 1;
-}
-
-message EntityIdRepeatedFieldValue {
-    repeated string value = 1;
-}
-
-// Unsupported ID types.
-
-message EntityIdByteStringFieldValue {
-    bytes value = 1;
-}
-
-message EntityIdDoubleFieldValue {
-    double value = 1;
+    google.protobuf.Any any = 1;
 }
