+++ base/src/test/proto/spine/test/validate/commands.proto
@@ -19,28 +19,46 @@
 //
 syntax = "proto3";
 
-package spine.test.validate.msg;
+package spine.test.validate;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
 option java_multiple_files = true;
-option java_outer_classname = "ValidationTestAltFieldsProto";
-option java_package = "io.spine.test.validate.msg.altfields";
+option java_outer_classname = "ValidationTestCommandsProto";
+option java_package = "io.spine.test.validate.commands";
+
+import "google/protobuf/wrappers.proto";
 
 
-message PersonName {
-    option (required_field) = "first_name | honorific_prefix & last_name";
+// Messages for entity ID in commands validation tests.
 
-    string honorific_prefix = 1;
-    string first_name = 2;
-    string last_name = 3;
-    string honorific_suffix = 4;
+message EntityIdMsgFieldValue {
+    google.protobuf.StringValue value = 1;
 }
 
-message MessageWithMissingField {
-    option (required_field) = "name | else";
+message EntityIdStringFieldValue {
+    string value = 1;
+}
 
-    string name = 1;
-    bool present = 2;
+message EntityIdIntFieldValue {
+    int32 value = 1;
+}
+
+message EntityIdLongFieldValue {
+    int64 value = 1;
+}
+
+message EntityIdRepeatedFieldValue {
+    repeated string value = 1;
+}
+
+// Unsupported ID types.
+
+message EntityIdByteStringFieldValue {
+    bytes value = 1;
+}
+
+message EntityIdDoubleFieldValue {
+    double value = 1;
 }
