+++ base/src/test/proto/spine/test/validate/messages.proto
@@ -19,14 +19,14 @@
 //
 syntax = "proto3";
 
-package spine.test.validate.msg;
+package spine.test.validate;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
 option java_multiple_files = true;
 option java_outer_classname = "ValidationTestMessagesProto";
-option java_package = "io.spine.test.validate.msg";
+option java_package = "io.spine.test.validate";
 
 import "google/protobuf/timestamp.proto";
 import "google/protobuf/wrappers.proto";
